Add files via upload
Project Description

The game starts with two players as active members, having a one- print board on the screen which displays 9 boxes (i.e., 3×3). The first person who starts the game has to enter either X or O for any one box, followed by the other player entering the other X or O (opposite to what the first player has entered). this continues unless any one of them gets a line cut either diagonally or straight. And the person who founds the line is the winner of the game.